TIC-110: Refactored complete order form.
